+++ src/libs/rustbase/src/arch/gem5/envdata.rs
@@ -31,6 +31,6 @@ pub struct EnvData {
 
 pub fn get() -> &'static mut EnvData {
     unsafe {
-        intrinsics::transmute(0x6000 as u64)
+        intrinsics::transmute(0x6000 as usize)
     }
 }
